+++ src/components/Inputs/FormSelect/FormSelectAsyncField.tsx
@@ -1,132 +1,11 @@
 import Autocomplete, { AutocompleteRenderInputParams } from '@mui/material/Autocomplete';
-import { Controller, useFormContext, FieldValues, Path, ControllerRenderProps } from 'react-hook-form';
-import { FormControl, TextField, Tooltip, CircularProgress, FormHelperText } from '@mui/material';
-import { get } from 'lodash';
+import { FieldValues, Path, ControllerRenderProps } from 'react-hook-form';
+import { TextField, CircularProgress } from '@mui/material';
 import { SyntheticEvent, UIEvent, useCallback, useEffect, useMemo, useState } from 'react';
 import { useDebounce } from 'use-debounce';
+import { Option } from '@Components/Inputs/FormMultiSelect/FormMultiSelect';
 
-interface Option {
-  value: number | string | undefined;
-  label: string;
-}
-
-// Define the props for the component
-export interface FormSelectProps<TFieldValues extends FieldValues> {
-  name: Path<TFieldValues>;
-  options?: Option[];
-  fetchOptions?: (
-    searchQuery: string,
-    page: number,
-  ) => Promise<{
-    data: Option[];
-    totalPages: number;
-  }>;
-  fetchOptionById?: (id: number) => Promise<{ data: Option }>;
-  label: string;
-  required?: boolean;
-  tooltip?: string;
-  id?: string;
-  disabled?: boolean;
-}
-
-export function FormSelect<TFieldValues extends FieldValues>({
-  name,
-  options,
-  fetchOptions,
-  fetchOptionById,
-  label,
-  required,
-  tooltip,
-  id,
-  disabled,
-}: FormSelectProps<TFieldValues>) {
-  const {
-    control,
-    formState: { errors },
-  } = useFormContext<TFieldValues>();
-
-  const errorMessage = get(errors, name)?.message;
-
-  const getOptionLabel = useCallback((option: Option) => option.label, []);
-
-  const renderTextFieldInput = useCallback(
-    (params: AutocompleteRenderInputParams) => {
-      return <TextField {...params} label={`${label}${required ? '*' : ''}`} />;
-    },
-    [label, required],
-  );
-
-  const onChange = useCallback(
-    (field: ControllerRenderProps<TFieldValues, Path<TFieldValues>>) =>
-      (_event: SyntheticEvent<Element, Event>, newValue: Option | null) => {
-        field.onChange(newValue ? newValue.value : null);
-      },
-    [],
-  );
-
-  const controllerRenderer = useCallback(
-    ({ field }: { field: ControllerRenderProps<TFieldValues, Path<TFieldValues>> }) => {
-      if (fetchOptions && fetchOptionById) {
-        return (
-          <AsyncField
-            field={field}
-            fetchOptions={fetchOptions}
-            fetchOptionById={fetchOptionById}
-            label={label}
-            required={required}
-            id={id}
-            disabled={disabled}
-          />
-        );
-      } else {
-        // Use synchronous options
-        const localOptions = options || [];
-        const selectedOption = localOptions.find((option) => option.value === field.value) || null;
-
-        return (
-          <Autocomplete
-            data-testid={`controlled-select-${name}`}
-            options={localOptions}
-            id={id}
-            getOptionLabel={getOptionLabel}
-            onChange={onChange(field)}
-            value={selectedOption}
-            renderInput={renderTextFieldInput}
-            disabled={disabled}
-          />
-        );
-      }
-    },
-    [
-      fetchOptions,
-      fetchOptionById,
-      label,
-      required,
-      id,
-      options,
-      name,
-      getOptionLabel,
-      onChange,
-      renderTextFieldInput,
-      disabled,
-    ],
-  );
-
-  return (
-    <FormControl sx={{ width: '100%' }} error={Boolean(errorMessage)}>
-      <Tooltip title={tooltip || ''} arrow enterDelay={2000} leaveDelay={200} placement="top" enterNextDelay={2000}>
-        <>
-          <Controller control={control} name={name} render={controllerRenderer} />
-          {Boolean(errorMessage) && (
-            <FormHelperText>{typeof errorMessage === 'string' ? errorMessage : undefined}</FormHelperText>
-          )}
-        </>
-      </Tooltip>
-    </FormControl>
-  );
-}
-
-interface AsyncFieldProps<TFieldValues extends FieldValues> {
+interface FormSelectAsyncFieldProps<TFieldValues extends FieldValues> {
   field: ControllerRenderProps<TFieldValues, Path<TFieldValues>>;
   label: string;
   required?: boolean;
@@ -143,7 +22,7 @@ interface AsyncFieldProps<TFieldValues extends FieldValues> {
   disabled?: boolean;
 }
 
-export const AsyncField = <T extends FieldValues>({
+export const FormSelectAsyncField = <T extends FieldValues>({
   field,
   label,
   required,
@@ -152,7 +31,7 @@ export const AsyncField = <T extends FieldValues>({
   onChangeOverride,
   id,
   disabled,
-}: AsyncFieldProps<T>) => {
+}: FormSelectAsyncFieldProps<T>) => {
   const [options, setOptions] = useState<Option[]>([]);
   const [inputValue, setInputValue] = useState('');
   const [loading, setLoading] = useState(false);
